Move form modals into its own folder.
