+++ js/src/components/shipping-rate-section/minimum-order-card/minimum-order-form-modals/edit-minimum-order-form-modal.js
@@ -12,11 +12,11 @@ import MinimumOrderFormModal from './minimum-order-form-modal';
 
 /**
  * @typedef { import(".~/data/actions").CountryCode } CountryCode
- * @typedef { import("./typedefs.js").MinimumOrderGroup } MinimumOrderGroup
+ * @typedef { import("../typedefs.js").MinimumOrderGroup } MinimumOrderGroup
  */
 
 /**
- * Display the add minimum order modal that is wrapped in a Form.
+ * Display the edit minimum order modal that is wrapped in a Form.
  *
  * When users submit the form, `props.onRequestClose` will be called first, and then followed by `props.onSubmit`.
  * If we were to call `props.onSubmit` first, it may cause some state change in the parent component and causing this component not to be rendered,
@@ -26,14 +26,21 @@ import MinimumOrderFormModal from './minimum-order-form-modal';
  * @param {Array<CountryCode>} props.countryOptions Array of country codes options, to be used as options in AppCountrySelect.
  * @param {MinimumOrderGroup} props.initialValues Initial values for the form.
  * @param {(values: MinimumOrderGroup) => void} props.onSubmit Callback when the form is submitted, with the form value.
+ * @param {() => void} props.onDelete Callback when delete button is clicked.
  * @param {() => void} props.onRequestClose Callback to close the modal.
  */
-const AddMinimumOrderFormModal = ( {
+const EditMinimumOrderFormModal = ( {
 	countryOptions,
 	initialValues,
 	onSubmit,
 	onRequestClose = noop,
+	onDelete = noop,
 } ) => {
+	const handleDeleteClick = () => {
+		onRequestClose();
+		onDelete();
+	};
+
 	return (
 		<MinimumOrderFormModal
 			countryOptions={ countryOptions }
@@ -41,19 +48,30 @@ const AddMinimumOrderFormModal = ( {
 			renderButtons={ ( formProps ) => {
 				const { isValidForm, handleSubmit } = formProps;
 
-				const handleAddClick = () => {
+				const handleUpdateClick = () => {
 					onRequestClose();
 					handleSubmit();
 				};
 
 				return [
 					<Button
+						key="delete"
+						isTertiary
+						isDestructive
+						onClick={ handleDeleteClick }
+					>
+						{ __( 'Delete', 'google-listings-and-ads' ) }
+					</Button>,
+					<Button
 						key="save"
 						isPrimary
 						disabled={ ! isValidForm }
-						onClick={ handleAddClick }
+						onClick={ handleUpdateClick }
 					>
-						{ __( 'Add minimum order', 'google-listings-and-ads' ) }
+						{ __(
+							'Update minimum order',
+							'google-listings-and-ads'
+						) }
 					</Button>,
 				];
 			} }
@@ -63,4 +81,4 @@ const AddMinimumOrderFormModal = ( {
 	);
 };
 
-export default AddMinimumOrderFormModal;
+export default EditMinimumOrderFormModal;
