+++ js/src/components/shipping-rate-section/minimum-order-card/minimum-order-form-modals/index.js
@@ -1,0 +1,2 @@
+export { default as AddMinimumOrderFormModal } from './add-minimum-order-form-modal';
+export { default as EditMinimumOrderFormModal } from './edit-minimum-order-form-modal';
